feat: CommonType 생성 -> 분리된 type 정리
-ghojeong/issue-tracker/#140
