+++ frontend/issue-cracker/src/components/common/sideBar/data/AssigneeData.tsx
@@ -3,9 +3,9 @@ import styled from 'styled-components';
 import { Issue as S } from '../../../styles/CommonStyles';
 import { v4 as uuidv4 } from 'uuid';
 import SideBarDropAssignee from '../SideBarDropAssignee';
-import { AssigneeDataProps } from '../../../../utils/types/sideBarType';
+import { AssigneeStateProps } from '../../../../utils/types/commonTypes';
 
-const AssigneeData = ({ userData }: AssigneeDataProps): JSX.Element => {
+const AssigneeData = ({ userData }: AssigneeStateProps): JSX.Element => {
   return (
     <>
       {userData?.map((user) => (
