+++ frontend/issue-cracker/src/components/layout/IssueDetail/IssueDetailBox/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
 import IssueDetailInput from './IssueDetailInput';
-import { IssueDataProps } from '../../../../utils/types/IssueDataType';
 import IssueDetailSidebar from '../IssueDetailBox/IssueDetailSidebar';
 
 const IssueDetailBox = (): JSX.Element => {
