+++ frontend/issue-cracker/src/components/layout/IssueList/IssueTable/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import { useRecoilValue } from 'recoil';
 import { issueListState } from '../../../../store/Recoil';
