--- frontend/issue-cracker/src/utils/types/milestoneDataType.ts
@@ -1,14 +1,0 @@
-import { IssueDataProps } from './IssueDataType';
-
-export interface MilestoneInfo {
-  description: string;
-  dueDate: string;
-  title: string;
-  status: string;
-}
-
-export interface MilestoneDataProps {
-  id: number;
-  issues: IssueDataProps[];
-  milestoneInfo: MilestoneInfo;
-}
